refactor: 💡 Add exceptions in a Video package
✅ Closes: #21
