+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/video/VideoInvalidTeacherImageURLException.java
@@ -1,4 +1,4 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.video;
 
 /**
  * Exception class for when the Video TeacherImageURL field is null
