+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/video/VideoInvalidTitleException.java
@@ -1,4 +1,4 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.video;
 
 /**
  * Exception class for when the Video Title field is null
