+++ apps/java-api/src/main/java/org/scoalaonline/api/exception/video/VideoNotFoundException.java
@@ -1,4 +1,4 @@
-package org.scoalaonline.api.exception;
+package org.scoalaonline.api.exception.video;
 
 /**
  * Exception class for when the Video entity cannot be found.
